+++ src/messages.ts
@@ -9,31 +9,23 @@ export const libString = (lib: string, ver?: string) => {
 	return `${lib}${vstr}`;
 };
 
-export const writeError = (library, version, reason) => {
-	return error(
-		library,
-		`${library} ${version} ${reason}`
-	);
-};
-
-export const writeMinVersionError = (library, version, minVersion, reason): Tip => {
+export const writeMinVersionError = (library: string, version: string, minVersion: string, reason: string): Tip => {
 	return new Tip(
 		library,
 		`${library} must be upgraded from ${version} to at least version ${minVersion}${reason ? ' ' + reason : ''}`,
 		TipType.Error, undefined,
-		`npm install ${library}@latest --save-exact`, 
-		`Upgrade`, 
+		`npm install ${library}@latest --save-exact`,
+		`Upgrade`,
 		`${library} successfully updated.`);
 };
-
 export const writeMinVersionWarning = (library: string, version: string, minVersion: string, reason: string, url?: string): Tip => {
 	let r = reason ? ' ' + reason : '';
 	if (url) r = `[${reason}](${url})`;
 	return new Tip(library,
 		`Update to at least ${minVersion}${reason ? ' ' + reason : ''}`, TipType.Idea,
 		`${library} ${version} should be updated to at least ${minVersion}${reason ? ' ' + reason : ''}`,
-		`npm install ${library}@latest --save-exact`, 
-		`Upgrade`, 
+		`npm install ${library}@latest --save-exact`,
+		`Upgrade`,
 		`${library} successfully updated.`);
 };
 
@@ -42,8 +34,8 @@ export const writeConsistentVersionWarning = (lib1: string, ver1: string, lib2: 
 		lib2,
 		`Version of ${libString(lib2, ver2)} should match ${libString(lib1, ver1)}`,
 		TipType.Error, undefined,
-		`npm install ${lib2}@${ver1} --save-exact`, 
-		`Upgrade`, 
+		`npm install ${lib2}@${ver1} --save-exact`,
+		`Upgrade`,
 		`${lib2} successfully updated.`);
 };
 
@@ -52,7 +44,7 @@ export const writeConsistentVersionError = (lib1: string, ver1: string, lib2: st
 		lib2,
 		`Version of ${libString(lib2, ver2)} must match ${libString(lib1, ver1)}`,
 		TipType.Error, undefined,
-		`npm install ${lib2}@${ver1} --save-exact`, 
-		`Upgrade`, 
+		`npm install ${lib2}@${ver1} --save-exact`,
+		`Upgrade`,
 		`${lib2} successfully updated.`);
 };